Reworked the gnutls code. Now more consistant with windows, and more generic. Still does not validate the server's certificate, so disabled by default.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4402 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/gl/gl_model.c
@@ -75,7 +75,7 @@ model_t *lightmodel;
 int numlightdata;
 qboolean writelitfile;
 
-int relitsurface;
+long relitsurface;
 void RMod_UpdateLightmap(int snum)
 {
 	msurface_t *s;
